fix(UserFilter): created seperate component for genericAutoComplete and fixed style
