+++ client/src/models/ActiveStatus.ts
@@ -1,0 +1,6 @@
+interface activeStatus {
+    displayName: string;
+    value: boolean;
+}
+
+export default activeStatus;
